Remove reused code on all reading pages
