+++ src/components/Reading/Reading.css
@@ -1,0 +1,11 @@
+div {
+  color: white;
+}
+
+section {
+  font-size: 25px;
+}
+
+.reading-card {
+  padding-top: 5vh;
+}
