+++ src/components/TodaysReading/TodaysReading.js
@@ -1,9 +1,10 @@
 import React, { Component } from "react";
 import { Link } from "react-router-dom";
 import PropTypes from 'prop-types';
-import "./ReadingPage.css";
+import Reading from "../Reading/Reading"
+import "./TodaysReading.css";
 
-class ReadingPage extends Component {
+class TodaysReading extends Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -38,34 +39,13 @@ class ReadingPage extends Component {
           <button className="tomorrow-btn">Tomorrows Reading→</button>
         </Link>
         <br />
-        <section className="reading-card">
-          Date: {this.state.responseReading.current_date} <br />
-          Description: {this.state.responseReading.description} <br />
-          Sign Compatibility: {this.state.responseReading.compatibility} <br />
-          Lucky Number: {this.state.responseReading.lucky_number} <br />
-          Lucky Time: {this.state.responseReading.lucky_time} <br />
-          Color: {this.state.responseReading.color} <br />
-          Date Range: {this.state.responseReading.date_range} <br />
-          Mood: {this.state.responseReading.mood} <br />
-          <button
-            className="save-btn"
-            onClick={() => {
-              localStorage.setItem(
-                this.state.responseReading.lucky_number +
-                  this.state.responseReading.lucky_time,
-                JSON.stringify(this.state.responseReading)
-              );
-            }}
-          >
-            Save Reading
-          </button>
-        </section>
+        <Reading reading={this.state.responseReading} />
       </div>
     );
   }
 }
 
-export default ReadingPage;
-ReadingPage.propTypes = {
+export default TodaysReading;
+TodaysReading.propTypes = {
   currentZodiac: PropTypes.string
 }
