+++ src/components/TomorrowsReading/TomorrowsReading.js
@@ -1,6 +1,7 @@
 import React, { Component } from "react";
 import { Link } from "react-router-dom";
 import PropTypes from 'prop-types';
+import Reading from "../Reading/Reading"
 import "./TomorrowsReading.css";
 
 class TomorrowsReading extends Component {
@@ -33,28 +34,7 @@ class TomorrowsReading extends Component {
           <button className="today-btn">← Todays reading</button>
         </Link>{" "}
         <br />
-        <section className="reading-card">
-          Date: {this.state.responseReading.current_date} <br />
-          Description: {this.state.responseReading.description} <br />
-          Compatibility: {this.state.responseReading.compatibility} <br />
-          Lucky Number: {this.state.responseReading.lucky_number} <br />
-          Lucky Time: {this.state.responseReading.lucky_time} <br />
-          Color: {this.state.responseReading.color} <br />
-          Date Range: {this.state.responseReading.date_range} <br />
-          Mood: {this.state.responseReading.mood} <br />
-          <button
-            className="save-btn"
-            onClick={() => {
-              localStorage.setItem(
-                this.state.responseReading.lucky_number +
-                  this.state.responseReading.lucky_time,
-                JSON.stringify(this.state.responseReading)
-              );
-            }}
-          >
-            Save Reading
-          </button>
-        </section>
+        <Reading reading={this.state.responseReading} />
       </div>
     );
   }
